fix enabling/disabling of all records when hover
